Add a build_info metric and a promu build spec.
This makes the exporter expose a dnsmasq_exporter_build_info metric, which not only allows you to see what version you're running, but also allows you to dynamically select the job label in a Gragana dashboard - making the dashboards more portable.

To make sure the the build_info metric is correctly populated, I've added a promu build spec, allowing the exporter to be build with promu, the tool we use to build Prometheus and various exporters.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>